refactor(services): update paths :recycle:
