+++ packages/services/README.md
@@ -1,9 +1,38 @@
 ## ディレクトリ構成
+ディレクトリの基本的な構成は以下の通りです。
 
 ```
-packages/services
- |
- ├─ src/
- |   ├─ common/                    # サービスに依存しない共通の機能
- |   └─ individuals/{対象サービス名} # どんなサービスにも使えるわけではないがある程度汎用的な機能 (ランキング, TODOリスト, etc)
+src
+├── subdomains  # それぞれ独立して成立するサービス
+│   └── user
+│       ├── Entities
+│       │   └── User.ts
+│       ├── Events
+│       │   └── UserCreated.ts
+│       ├── Infrastructures
+│       │   └── typeorm
+│       │       └── User.ts
+│       ├── Mappers
+│       │   └── UserMap.ts
+│       ├── Repositories
+│       │   ├── UserRepository.ts
+│       │   └── implements
+│       │       └── TypeOrmUserRepository.ts
+│       ├── Subscriptions
+│       │   └── AfterUserCreated.ts
+│       ├── UseCases
+│       │   └── CreateUser
+│       │       ├── CreateUserUseCase.ts
+│       │       ├── CreateUserDTO.ts
+│       │       ├── CreateUserErrors.ts
+│       │       └── CreateUserController.ts
+│       └── ValueObjects
+│           └── UserName.ts
+│
+└── shared  # 共通のコアロジックや、ベースクラスなどを配置
+    ├── core
+    │   └── xxx.ts
+    └── domain
+        └── ValueObject # 値オブジェクトを定義
+            └── Id.ts
 ```
