+++ packages/services/src/index.ts
@@ -1,0 +1,4 @@
+/**
+ * 共通のコアロジックや、ベースクラス
+ */
+export { Shared } from "~/shared";
