+++ packages/services/src/shared/core/ValueObject.ts
